Replace singleton code with `LLMS_Trait_Singleton`.
