+++ class-lifterlms.php
@@ -23,8 +23,13 @@ defined( 'ABSPATH' ) || exit;
  * @since 4.0.0 Update session management.
  *              Remove deprecated class files and variables.
  *              Move includes (file loading) into the LLMS_Loader class.
+ * @since [version] Replace singleton code with `LLMS_Trait_Singleton`.
+ *
+ * @method static LifterLMS instance()
  */
-final class LifterLMS {
+final class LifterLMS  {
+
+	use LLMS_Trait_Singleton;
 
 	/**
 	 * LifterLMS Plugin Version.
@@ -32,13 +37,6 @@ final class LifterLMS {
 	 * @var string
 	 */
 	public $version = '5.2.0';
-
-	/**
-	 * Singleton instance of LifterLMS.
-	 *
-	 * @var LifterLMS
-	 */
-	protected static $_instance = null;
 
 	/**
 	 * LLMS_Assets instance
@@ -62,22 +60,6 @@ final class LifterLMS {
 	public $session = null;
 
 	/**
-	 * Main Instance of LifterLMS
-	 * Ensures only one instance of LifterLMS is loaded or can be loaded.
-	 *
-	 * @see      LLMS()
-	 * @return   LifterLMS - Main instance
-	 * @since    1.0.0
-	 * @version  1.0.0
-	 */
-	public static function instance() {
-		if ( is_null( self::$_instance ) ) {
-			self::$_instance = new self();
-		}
-		return self::$_instance;
-	}
-
-	/**
 	 * LifterLMS Constructor.
 	 *
 	 * @since 1.0.0
@@ -85,6 +67,7 @@ final class LifterLMS {
 	 * @since 4.0.0 Load `$this->session` at `plugins_loaded` in favor of during class construction.
 	 *               Remove deprecated `__autoload()` & initialize new file loader class.
 	 * @since 4.13.0 Check site duplicate status on `admin_init`.
+	 * @since [version] Move the loading of the LifterLMS autoloader to the main `lifterlms.php` file.
 	 *
 	 * @return void
 	 */
@@ -98,8 +81,6 @@ final class LifterLMS {
 		 * so that our custom "safe" location will always load first.
 		 */
 		$this->localize();
-
-		require_once LLMS_PLUGIN_DIR . 'includes/class-llms-loader.php';
 
 		$this->define_constants();
