+++ includes/class.llms.emails.php
@@ -17,8 +17,13 @@ defined( 'ABSPATH' ) || exit;
  *
  * @since 1.0.0
  * @since 3.8.0 Unknown.
+ * @since [version] Replace singleton code with `LLMS_Trait_Singleton`.
+ *
+ * @method static LLMS_Emails instance()
  */
 class LLMS_Emails {
+
+	use LLMS_Trait_Singleton;
 
 	/**
 	 * Class names of all emails
@@ -26,27 +31,6 @@ class LLMS_Emails {
 	 * @var string[]
 	 */
 	public $emails;
-
-	/**
-	 * protected private instance of email
-	 *
-	 * @var LLMS_Emails
-	 */
-	protected static $_instance = null;
-
-	/**
-	 * Create instance of class
-	 *
-	 * @since 1.0.0
-	 *
-	 * @return LLMS_Emails
-	 */
-	public static function instance() {
-		if ( is_null( self::$_instance ) ) {
-			self::$_instance = new self();
-		}
-		return self::$_instance;
-	}
 
 	/**
 	 * Constructor
