+++ includes/class.llms.engagements.php
@@ -18,8 +18,13 @@ defined( 'ABSPATH' ) || exit;
  * @since 2.3.0
  * @since 3.30.3 Fixed spelling errors.
  * @since 3.39.0 Added `llms_rest_student_registered` as action hook.
+ * @since [version] Replace singleton code with `LLMS_Trait_Singleton`.
+ *
+ * @method static LLMS_Engagements instance()
  */
 class LLMS_Engagements {
+
+	use LLMS_Trait_Singleton;
 
 	/**
 	 * Enable debug logging
@@ -28,25 +33,6 @@ class LLMS_Engagements {
 	 * @var boolean
 	 */
 	private $debug = false;
-
-	/**
-	 * Protected instance of class
-	 *
-	 * @var LLMS_Engagements
-	 */
-	protected static $_instance = null;
-
-	/**
-	 * Create instance of class
-	 *
-	 * @return LLMS_Engagements Instance of engagements class.
-	 */
-	public static function instance() {
-		if ( is_null( self::$_instance ) ) {
-			self::$_instance = new self();
-		}
-		return self::$_instance;
-	}
 
 	/**
 	 * Constructor
