+++ includes/class.llms.payment.gateways.php
@@ -15,8 +15,13 @@ defined( 'ABSPATH' ) || exit;
  *
  * @since 1.0.0
  * @since 3.0.0 Unknown.
+ * @since [version] Replace singleton code with `LLMS_Trait_Singleton`.
+ *
+ * @method static LLMS_Payment_Gateways instance()
  */
 class LLMS_Payment_Gateways {
+
+	use LLMS_Trait_Singleton;
 
 	/**
 	 * Payment Gateways
@@ -24,28 +29,6 @@ class LLMS_Payment_Gateways {
 	 * @var array
 	 */
 	public $payment_gateways = array();
-
-	/**
-	 * Private instance of class
-	 *
-	 * @var null
-	 */
-	private static $_instance = null;
-
-	/**
-	 * Create instance of class
-	 *
-	 * @return self
-	 */
-	public static function instance() {
-
-		if ( is_null( self::$_instance ) ) {
-			self::$_instance = new self();
-		}
-
-		return self::$_instance;
-
-	}
 
 	/**
 	 * Constructor
