+++ tests/phpunit/unit-tests/class-llms-test-main-class.php
@@ -27,15 +27,14 @@ class LLMS_Test_Main_Class extends LLMS_UnitTestCase {
 	}
 
 	/**
-	 * test the _instance variable
+	 * Test the `instance` property.
 	 *
 	 * @since 3.3.1
-	 *
-	 * @return void
+	 * @since [version] Rename `_instance` property to `instance`.
 	 */
 	public function test_llms_instance() {
 
-		$this->assertClassHasStaticAttribute( '_instance', 'LifterLMS' );
+		$this->assertClassHasStaticAttribute( 'instance', 'LifterLMS' );
 
 	}
